+++ scripts/changelog.ts
@@ -537,7 +537,9 @@ useScopedSelector(state => {
         '**plugin**. The return type of the built-in `upload` state type is no longer callable. Replace `state()` with `state.get()` or `state.value`.'
       ],
       added: [
-        '**plugin**. The return type of the built-in scalar state types (i.e. `boolean`, `number`, `string`, `scalar`, `serializedScalar`) now allows to set the value. Instead of `state.set("foo")`, you may now also use `state.value = "foo"`.'
+        '**plugin**. The return type of the built-in scalar state types (i.e. `boolean`, `number`, `string`, `scalar`, `serializedScalar`) now allows to set the value. Instead of `state.set("foo")`, you may now also use `state.value = "foo"`.',
+        '**store**. Add new selector `hasFocusedChild` that checks whether the given document has any (direct) child that is focused',
+        '**store**. Add new selector `hasFocusedDescendant` that checks whether the given document has any descendant that is focused (i.e. also recursively checks children of children)'
       ]
     }
   ])